requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
lancedb>=0.3.0
sentence-transformers>=2.2.0

# Model and inference
# sgllm  # Uncomment when available

# Static analysis tools
ruff>=0.1.0
# pyright  # Install via npm: npm install -g pyright
# eslint   # Install via npm: npm install -g eslint

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Monitoring and logging
psutil>=5.9.0

# Optional: Dependency injection
dependency-injector>=4.41.0

# Optional: Async support
aiofiles>=23.0.0
asyncio-mqtt>=0.13.0